+++ examples/create-trade-towards-allocation.rs
@@ -220,6 +220,9 @@ async fn main() -> anyhow::Result<()> {
         println!(" {trade_url}");
     }
     println!("Done");
+
+    montrose.disconnect().await;
+
     Ok(())
 }