+++ examples/show-data.rs
@@ -91,5 +91,7 @@ async fn main() -> anyhow::Result<()> {
         }
     }
 
+    montrose.disconnect().await;
+
     Ok(())
 }